Migrate to the new probe file access system
